separate the Timers from main to be able to use named Errors in Timers
